split templates
